+++ modules/test/cobertura-agent/src/com/liferay/cobertura/agent/InstrumentationAgent.java
@@ -14,6 +14,8 @@
 
 package com.liferay.cobertura.agent;
 
+import com.liferay.cobertura.agent.instrument.ProjectDataUtil;
+import com.liferay.cobertura.agent.instrument.CoberturaClassFileTransformer;
 import java.io.File;
 import java.io.IOException;
