+++ modules/test/cobertura-agent/src/com/liferay/cobertura/agent/instrument/CoberturaClassFileTransformer.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.cobertura.agent;
+package com.liferay.cobertura.agent.instrument;
 
+import com.liferay.cobertura.agent.InstrumentationAgent;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.FileWriter;
