+++ modules/test/cobertura-agent/src/com/liferay/cobertura/agent/instrument/ProjectDataUtil.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.cobertura.agent;
+package com.liferay.cobertura.agent.instrument;
 
+import com.liferay.cobertura.agent.InstrumentationAgent;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileDescriptor;
